通过更多cpu tests

+++ npc/csrc/difftests/memory.h
@@ -2,8 +2,8 @@ typedef uint64_t paddr_t;
 typedef uint64_t vaddr_t;
 
 uint8_t *guest_to_host(paddr_t paddr);
-uint64_t pmem_read(paddr_t addr, int len);
-void pmem_write(paddr_t addr, int len, uint64_t data);
+uint64_t pmem_read(paddr_t addr, size_t len);
+void pmem_write(paddr_t addr, size_t len, uint64_t data);
 
 extern uint8_t pmem[];
 long load_image(char *filename);
